support log and masked_fill multicore
1. support log multicore
2. support masked_fill multicore
3. support masked_fill broadcast cases corresponding to backend

Known issues:
1. masked_fill input and mask different dims not tested
2. masked_fill large shape has problems, need to modify backend.

+++ python/test/mask_fill.py
@@ -21,7 +21,7 @@ def case1():
 
     TPU = "privateuseone:0"
     mask_tpu = torch.full((tgt_len, tgt_len), torch.tensor(torch.finfo(dtype).min, device=TPU), device=TPU)
-    mask_cond_tpu = torch.arange(mask_tpu.size(-1), device=TPU)
+    mask_cond_tpu = torch.arange(mask_tpu.size(-1), dtype=dtype, device=TPU)
     s1_t = (mask_cond_tpu + 1).view(mask_tpu.size(-1),1)
     s2_t = mask_cond_tpu < s1_t
     mask_tpu.masked_fill_(s2_t, 0)
